refactor(lib,contracts): extract shared admin/validation/types/storage keys and migrate core contracts to DRY utilities
